+++ amundsen_application/static/css/_typography-default.scss
@@ -119,7 +119,7 @@ body {
   font-weight: $font-weight-body-bold;
 }
 
-.column-type {
+.column-name {
   color: $brand-color-3;
   font-size: 13px;
   font-family: $font-family-monospace;
