+++ amundsen_application/static/js/components/TableDetail/ColumnList/styles.scss
@@ -1,0 +1,3 @@
+.column-list {
+  margin: 0;
+}
