+++ amundsen_application/static/js/components/TableDetail/ColumnListItem/styles.scss
@@ -1,0 +1,107 @@
+@import 'variables';
+
+.list-group-item .column-list-item {
+  display: flex;
+  padding: 16px 24px;
+
+
+  .column-details {
+    flex-basis: 50%;
+    padding-right: 16px;
+  }
+  .resource-type {
+
+  }
+  .badges {
+
+  }
+  .actions {
+
+  }
+
+
+
+
+
+  //text-decoration: none;
+  //display: flex;
+  //flex-direction: column;
+  ///* tmp fixes until we refactor/settle on styles */
+  //cursor: default;
+  //border-top-color: $gray-lighter !important;
+  //border-bottom-color: $gray-lighter !important;
+  //background-color: transparent !important;
+  //padding: 10px 4px;
+  //
+  //.description {
+  //  max-width: 100%;
+  //  min-width: 0;
+  //  padding-right: 32px;
+  //}
+  //
+  //// TODO - move to common styles
+  //.truncated,
+  //.truncated .editable-text {
+  //  white-space: nowrap;
+  //  overflow: hidden;
+  //  text-overflow: ellipsis;
+  //}
+  //
+  //.column-info {
+  //  display: flex;
+  //  flex-direction: row;
+  //  margin: -10px -4px 0;
+  //  padding: 10px 4px 0;
+  //
+  //  .title-section {
+  //    width: 100%;
+  //    overflow-wrap: break-word;
+  //
+  //    .title-row {
+  //      display: flex;
+  //      flex-direction: row;
+  //
+  //      .name {
+  //        margin-right: 8px;
+  //        max-width: 85%;
+  //      }
+  //      .column-type {
+  //        margin-top: 4px;
+  //        max-width: 15%;
+  //      }
+  //    }
+  //  }
+  //
+  //  &.expandable {
+  //    cursor: pointer;
+  //
+  //    &:hover {
+  //      background-image: linear-gradient($gray-lighter, $gray-lighter, white);
+  //
+  //      .icon {
+  //        background-color: $brand-color-4;
+  //      }
+  //    }
+  //  }
+  //}
+
+  .column-stats {
+    display: block;
+    margin-top: 8px;
+    width: 100%;
+
+    .column-stat {
+      display: inline-block;
+      margin-top: 8px;
+      max-width: 150px;
+      min-width: 120px;
+      width: 20%;
+    }
+
+    .stat-collection-info {
+      color: $text-medium;
+      font-style: italic;
+      margin-top: 4px;
+    }
+  }
+}
